# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Mini ERP Backend API doc",
        "title": "Mini ERP Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/stock/adjust": {
            "post": {
                "description": "Adjust stock for a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockTransaction"
                ],
                "summary": "Adjust Stock",
                "parameters": [
                    {
                        "description": "Stock Adjust Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/command.StockAdjustRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/command.StockAdjustResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid input, reason required, or quantity cannot be zero",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found: Product does not exist",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stock/in": {
            "post": {
                "description": "Handle stock in for a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockTransaction"
                ],
                "summary": "Stock In",
                "parameters": [
                    {
                        "description": "Stock In Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/command.StockInRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/command.StockInResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid input or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found: Product does not exist",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\"s",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stock/out": {
            "post": {
                "description": "Handle stock out for a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StockTransaction"
                ],
                "summary": "Stock Out",
                "parameters": [
                    {
                        "description": "Stock Out Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/command.StockOutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/command.StockOutResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid input or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found: Product does not exist",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/suppliers/{id}": {
            "get": {
                "description": "Get supplier details by supplier ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Get a supplier by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Supplier ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Supplier"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update supplier information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Update a supplier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Supplier ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated supplier information",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/command.UpdateSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a supplier by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete a supplier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Supplier ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "command.StockAdjustRequest": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "ผู้ทำรายการ",
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "description": "+ เพิ่ม, - ลด",
                    "type": "integer"
                },
                "reason": {
                    "description": "REQUIRED สำหรับ ADJUST",
                    "type": "string"
                }
            }
        },
...
        "command.StockAdjustResult": {
            "type": "object",
            "properties": {
                "current_stock": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "transaction": {
                    "$ref": "#/definitions/model.StockTransaction"
                }
            }
        },
        "command.StockInRequest": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "ผู้ทำรายการ",
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                }
            }
        },
        "command.StockInResult": {
            "type": "object",
            "properties": {
                "current_stock": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "transaction": {
                    "$ref": "#/definitions/model.StockTransaction"
                }
            }
        },
        "command.StockOutRequest": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "ผู้ทำรายการ",
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "command.StockOutResult": {
            "type": "object",
            "properties": {
                "current_stock": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "transaction": {
                    "$ref": "#/definitions/model.StockTransaction"
                }
            }
        },
...
        "command.UpdateSupplierRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "model.Category": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "category_id": {
                    "type": "string"
                },
                "cost_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "product_code": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "selling_price": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.PurchaseOrder": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "TotalAmount     uint64              `gorm:\"not null\" json:\"total_amount\"`",
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "purchase_order_id": {
                    "type": "string"
                },
                "purchase_order_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PurchaseOrderItem"
                    }
                },
                "status": {
                    "$ref": "#/definitions/model.PurchaseOrderStatus"
                },
                "stock_transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StockTransaction"
                    }
                },
                "supplier": {
                    "$ref": "#/definitions/model.Supplier"
                },
                "supplier_id": {
                    "type": "string"
                }
            }
        },
        "model.PurchaseOrderItem": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "product_id": {
                    "type": "string"
                },
                "purchase_order_id": {
                    "type": "string"
                },
                "purchase_order_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "model.PurchaseOrderStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "CONFIRMED",
                "RECEIVED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "Draft",
                "Confirmed",
                "Received",
                "Cancelled"
            ]
        },
        "model.StockTransaction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "stock_transaction_id": {
                    "type": "string"
                },
                "type": {
                    "description": "e.g., \"IN\" or \"OUT\" or \"ADJUST\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TransactionType"
                        }
                    ]
                }
            }
        },
        "model.Supplier": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "purchaseOrders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PurchaseOrder"
                    }
                },
                "supplier_id": {
                    "type": "string"
                }
            }
        },
        "model.TransactionType": {
            "type": "string",
            "enum": [
                "IN",
                "OUT",
                "ADJUST"
            ],
            "x-enum-varnames": [
                "TransactionTypeIn",
                "TransactionTypeOut",
                "TransactionTypeAdjust"
            ]
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Example of \"Value\": Bearer \u003cyour_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}